Task Spring Boot part 2. Made methods User/BookServiceImplTemplate (JdbcTemplate).

+++ src/main/java/com/edu/ulab/app/service/impl/UserServiceImpl.java
@@ -52,7 +52,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
-        log.info("Deleted user with id=: {}", id);
+        if(getUserById(id)!=null){
+            userRepository.deleteById(id);
+            //log.info("Deleted user with id=: {}", id);
+        }
     }
 }
